[BUGFIX] Internally store boolean properties as integers
This avoids problems when persisting models with the ConnectionPool.

+++ Classes/OldModel/AbstractModel.php
@@ -340,7 +340,7 @@ abstract class AbstractModel extends \Tx_Oelib_TemplateHelper
     {
         $this->assertNonEmptyKey($key);
 
-        $this->recordData[$key] = (bool)$value;
+        $this->recordData[$key] = (int)(bool)$value;
     }
 
     /**
